Enhance security documentation and normalize notifications schema

+++ backend/src/main/java/com/smartcampus/backend/dto/notification/UnreadCountResponse.java
@@ -1,3 +1,3 @@
 package com.smartcampus.backend.dto.notification;
 
-public record UnreadCountResponse(long count) {}
+public record UnreadCountResponse(int count) {}
